requirements.txt
google-generativeai
python-dotenv
pymupdf
pillow
pyodbc
flask
flask-cors
